Não permite remoção de autores com notícias vinculadas

+++ app/Http/Controllers/Api/AuthorController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\AuthorRequest;
+use App\Models\Article;
 use App\Models\Author;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class AuthorController extends Controller
 {
@@ -43,6 +45,16 @@ class AuthorController extends Controller
 
     public function destroy(Author $author)
     {
+        $hasArticleForThisAuthor = Article::where('author_id', $author->id)->exists();
+
+        if ($hasArticleForThisAuthor) {
+            return response()->json([
+                'success' => false,
+                'status' => Response::HTTP_CONFLICT,
+                'message' => 'Não foi possível remover autor. Existem notícias vínculadas a ele.'
+            ]);
+        }
+    
         $author->delete();
         return response()->noContent();
     }
